refactor: add useChat hooks
